fix : drop 리렌더 최적화 완료
